minecraft: keep the resource pack transfer sizing with its state
The download constants and the chunk arithmetic sat in a file of their own,
but the state they size — downloadingPack, its chunk count and its requested
set — already lives with the queue, and two of the three symbols had a single
caller. Move them next to the types they describe.
