+++ minecraft/resource_pack_queue.go
@@ -15,6 +15,15 @@ import (
 // larger than 10 MiB outright, so any larger framing would have to stay under that.
 const DefaultResourcePackChunkSize = 1024 * 100
 
+// DefaultResourcePackMaxInFlightChunks is how many chunk requests a download keeps outstanding,
+// matching the vanilla client request window.
+const DefaultResourcePackMaxInFlightChunks = 100
+
+// maxResourcePackPrealloc bounds the buffer reserved up front for a pack. The size comes from
+// ResourcePacksInfo and is only a claim at that point, so a server would otherwise turn one small packet
+// into an allocation of any size it names. Larger packs still grow their buffer as chunks arrive.
+const maxResourcePackPrealloc = 32 << 20
+
 // resourcePackQueue is used to aid in the handling of resource pack queueing and downloading. Only one
 // resource pack is downloaded at a time.
 type resourcePackQueue struct {
@@ -113,4 +122,21 @@ func (queue *resourcePackQueue) NextPack() (pk *packet.ResourcePackDataInfo, ok 
 // AllDownloaded checks if all resource packs in the queue are downloaded.
 func (queue *resourcePackQueue) AllDownloaded() bool {
 	return len(queue.packsToDownload) == 0
+}
+
+// resourcePackChunkCount returns the number of chunks a pack of size bytes is split into. It reports false if
+// chunkSize is zero or if the resulting indices cannot be represented by ResourcePackChunkRequest.ChunkIndex,
+// which is a signed int32.
+func resourcePackChunkCount(size uint64, chunkSize uint32) (uint32, bool) {
+	if chunkSize == 0 {
+		return 0, false
+	}
+	count := size / uint64(chunkSize)
+	if size%uint64(chunkSize) != 0 {
+		count++
+	}
+	if count > uint64(1)<<31 {
+		return 0, false
+	}
+	return uint32(count), true
 }
